Testing the ReturnDate and the RentalFee
Testing the ReturnDate:
The purpose of this test:

This test is structured for the purpose of checking if the time
difference between test code and execution code.

1. const res = await exec();
- DB stores the data due to the execution of exec function

- Then, in real code we manually change the dateReturned property in one
of rental collection's documents.

- As test code is not able to recognize this change, we add distraction
calculation between two different time values.

- We don't need to get the exact calculated value from the distraction,
but what we need is just to check if the change of the value of dateReturned value in real code is reflected on Database or not.

- Sum up: Check if the change in time is successfully reflected on DB